* 'autorestv3' of https://github.com/Azure/autorest.python:
  Declare kwargs reserved (#525)
  updating to m4 4.12.276 (#524)

+++ test/azure/Expected/AcceptanceTests/AzureSpecials/azurespecialproperties/models/_models.py
@@ -20,7 +20,7 @@ class Error(msrest.serialization.Model):
     :param status:
     :type status: int
     :ivar constant_id: Required.  Default value: "1".
-    :vartype constant_id: float
+    :vartype constant_id: int
     :param message:
     :type message: str
     """
@@ -31,7 +31,7 @@ class Error(msrest.serialization.Model):
 
     _attribute_map = {
         'status': {'key': 'status', 'type': 'int'},
-        'constant_id': {'key': 'constantId', 'type': 'float'},
+        'constant_id': {'key': 'constantId', 'type': 'int'},
         'message': {'key': 'message', 'type': 'str'},
     }
