+++ test/azure/Expected/AcceptanceTests/AzureSpecials/azurespecialproperties/models/_models_py3.py
@@ -22,7 +22,7 @@ class Error(msrest.serialization.Model):
     :param status:
     :type status: int
     :ivar constant_id: Required.  Default value: "1".
-    :vartype constant_id: float
+    :vartype constant_id: int
     :param message:
     :type message: str
     """
@@ -33,7 +33,7 @@ class Error(msrest.serialization.Model):
 
     _attribute_map = {
         'status': {'key': 'status', 'type': 'int'},
-        'constant_id': {'key': 'constantId', 'type': 'float'},
+        'constant_id': {'key': 'constantId', 'type': 'int'},
         'message': {'key': 'message', 'type': 'str'},
     }
